refactor： refractor code
