+++ frontend/src/pages/TransportWorker/components/InventoryListCard.tsx
